- added modal confirmation dialog - added italian language

+++ src/app/post-list-item/post-list-item.component.ts
@@ -1,3 +1,5 @@
+import { ConfirmDialogComponent } from './../confirm-dialog/confirm-dialog.component';
+import { ModalService } from './../services/modal.service';
 import { PostService } from './../services/post.service';
 import { Post } from './../model/post';
 import { Component, OnInit, Input } from '@angular/core';
@@ -11,9 +13,18 @@ import { TranslateService } from '@ngx-translate/core';
 export class PostListItemComponent implements OnInit {
 
   @Input() post: Post;
+
+  confirmDialogInputs = {
+    titleKey: 'post-item.confirm-dialog.title',
+    messageKey: 'post-item.confirm-dialog.message',
+    confirmKey: 'post-item.confirm-dialog.confirm',
+    cancelKey: 'post-item.confirm-dialog.cancel'
+  }
+
   constructor(
     private postService: PostService,
-    translate: TranslateService
+    private modalService: ModalService,
+    public translate: TranslateService
     ) {}
 
   ngOnInit() {}
@@ -27,6 +38,17 @@ export class PostListItemComponent implements OnInit {
   }
 
   onRemove() {
-    this.postService.removePost(this.post);
+    const confirmDialogComponentRef = this.modalService.init(ConfirmDialogComponent, this.confirmDialogInputs, {});
+    
+    confirmDialogComponentRef.instance.confirmedEvent.subscribe(
+      (event) => this.onConfirmRemove(event)
+    );
+  }
+
+  onConfirmRemove(confirmed: boolean) {
+    if(confirm) {
+      this.postService.removePost(this.post);
+    }
+    this.modalService.destroy();
   }
 }
